+++ internal/checks/injection_coverage_85_test.go
@@ -7,7 +7,10 @@ import (
 	"path/filepath"
 	"strings"
 	"testing"
+
 	"time"
+
+	"github.com/pidginhost/csm/internal/platform"
 
 	"github.com/pidginhost/csm/internal/alert"
 	"github.com/pidginhost/csm/internal/config"
@@ -178,6 +181,7 @@ func TestAuditMailRootForwardEmpty(t *testing.T) {
 }
 
 func TestAuditMailSecureAuthDisabled(t *testing.T) {
+	withMockMTA(t, platform.MTAExim)
 	withMockOS(t, &mockOS{
 		stat: func(name string) (os.FileInfo, error) {
 			return nil, os.ErrNotExist
@@ -202,6 +206,7 @@ func TestAuditMailSecureAuthDisabled(t *testing.T) {
 }
 
 func TestAuditMailSecureAuthEnabled(t *testing.T) {
+	withMockMTA(t, platform.MTAExim)
 	withMockOS(t, &mockOS{
 		stat: func(name string) (os.FileInfo, error) {
 			return nil, os.ErrNotExist